backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21

# Database & ORM
sqlalchemy==2.0.45
alembic==1.17.2
psycopg2-binary==2.9.11

# Authentication & Security
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
pydantic==2.12.5
pydantic-settings==2.12.0

# Environment & Config
python-dotenv==1.2.1

# External APIs
anthropic==0.75.0
cloudinary==1.44.1
httpx==0.28.1

# File Processing
openpyxl==3.1.5
python-pptx==1.0.2
pdf2image==1.17.0
pillow==12.1.0
python-pptx==1.0.2

# CORS
fastapi-cors==0.0.6

# Logging & Monitoring
python-json-logger==4.0.0

# Development & Testing
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
httpx[http2]==0.28.1

# Code Quality
black==25.12.0
flake8==7.3.0
mypy==1.19.1
isort==7.0.0

# Type Stubs
types-python-jose==3.5.0.20250531
sqlalchemy-stubs==0.4

# Optional: Background Jobs (for future)
# celery==5.3.4
# redis==5.0.1

# Optional: Data Validation
email-validator==2.3.0